# packages/contracts/generated/openapi/platform.openapi.json
{
  "components": {
    "parameters": {
      "clientBuild": {
        "description": "Client build identifier",
        "in": "header",
        "name": "X-Client-Build",
        "required": true,
        "schema": {
          "maxLength": 80,
          "minLength": 1,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._+-]{0,79}$",
          "type": "string"
        }
      },
...
      "installationId": {
        "description": "Stable installation identifier",
        "in": "header",
        "name": "X-Client-Installation-Id",
        "required": true,
        "schema": {
          "maxLength": 128,
          "minLength": 1,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:+-]{0,127}$",
          "type": "string"
        }
      },
...
      "roleContextId": {
        "description": "Current authorized role-context identifier",
        "in": "header",
        "name": "X-Role-Context-Id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
      "schemaVersion": {
        "description": "Supported Milestone 1 contract schema version",
        "in": "header",
        "name": "X-Client-Schema-Version",
        "required": true,
        "schema": {
          "const": "1",
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "MpQueryContextResponse": {
        "additionalProperties": false,
        "properties": {
          "activeRelease": {
            "type": "null"
          },
          "availableMetricKeys": {
            "items": {
              "not": {}
            },
            "maxItems": 0,
            "type": "array"
          },
          "code": {
            "const": "DEPENDENCY_UNAVAILABLE",
            "type": "string"
          },
          "state": {
            "const": "UNAVAILABLE",
            "type": "string"
          }
        },
        "required": [
          "state",
          "code",
          "availableMetricKeys",
          "activeRelease"
        ],
        "type": "object",
        "x-data-classification": "C1"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "AUTHENTICATION_REQUIRED",
              "AUTHORIZATION_DENIED",
              "MFA_REQUIRED",
              "AUTHORIZATION_VERSION_CHANGED",
              "CONSENT_OR_ACCESS_VERSION_CHANGED",
              "DEVICE_BINDING_MISMATCH",
              "IDEMPOTENCY_KEY_REUSED",
              "EXPECTED_REVISION_MISMATCH",
              "INVALID_STATE_TRANSITION",
              "TOMBSTONED_ENTITY",
              "SOURCE_VERSION_EXPIRED",
              "EVIDENCE_INSUFFICIENT",
              "SYNC_CURSOR_INVALID",
              "SYNC_CURSOR_EXPIRED",
              "SYNC_BOOTSTRAP_REQUIRED",
              "SYNC_SCHEMA_UNSUPPORTED",
              "SYNC_BATCH_ID_REUSED",
              "CAUSAL_DEPENDENCY_UNSATISFIED",
              "ASSIGNMENT_CHANGED",
              "CLOCK_UNTRUSTED",
              "MEDIA_INTEGRITY_MISMATCH",
              "MEDIA_NOT_VERIFIED",
              "UPLOAD_INTENT_EXPIRED",
              "VOICE_PROPOSAL_EXPIRED",
              "VOICE_PROPOSAL_HASH_MISMATCH",
              "VISUAL_REVIEW_REQUIRED",
              "RELEASE_INVALIDATED",
              "RELEASE_UNAVAILABLE",
              "DEPENDENCY_UNAVAILABLE",
              "FILTER_NOT_ALLOWLISTED",
              "COMPARISON_NOT_RELEASABLE",
              "BATCH_ID_PAYLOAD_MISMATCH",
              "RATE_LIMITED"
            ],
            "type": "string"
          },
          "correlationId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "detail": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "fieldErrors": {
            "default": [],
            "items": {
              "additionalProperties": false,
              "properties": {
                "code": {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                "field": {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": [
                "field",
                "code"
              ],
              "type": "object"
            },
            "maxItems": 25,
            "type": "array"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "correlationId",
          "retryable",
          "fieldErrors"
        ],
        "type": "object",
        "x-data-classification": "C1"
      },
...
    },
    "securitySchemes": {
      "appCheck": {
        "in": "header",
        "name": "X-Firebase-AppCheck",
        "type": "apiKey"
      },
      "bearerAuth": {
        "bearerFormat": "Firebase ID token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated Milestone 1 contract. JSON fields use lower camel case.",
    "title": "Smart Fasal platform API",
    "version": "1.0.0-m1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/mp/query-context": {
      "get": {
        "operationId": "getMpQueryContext",
        "parameters": [
          {
            "$ref": "#/components/parameters/installationId"
          },
          {
            "$ref": "#/components/parameters/clientBuild"
          },
          {
            "$ref": "#/components/parameters/schemaVersion"
          },
          {
            "$ref": "#/components/parameters/roleContextId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpQueryContextResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request header, path or body failed schema or value validation",
            "x-smart-fasal-error-family": "REQUEST_VALIDATION",
            "x-smart-fasal-validation-sources": [
              "headers"
            ]
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Typed request failure",
            "x-smart-fasal-problem-codes": [
              "AUTHENTICATION_REQUIRED"
            ]
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Typed request failure",
            "x-smart-fasal-problem-codes": [
              "AUTHORIZATION_DENIED",
              "MFA_REQUIRED"
            ]
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Typed request failure",
            "x-smart-fasal-problem-codes": [
              "AUTHORIZATION_VERSION_CHANGED"
            ]
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Typed request failure",
            "x-smart-fasal-problem-codes": [
              "DEPENDENCY_UNAVAILABLE"
            ]
          }
        },
        "security": [
          {
            "appCheck": [],
            "bearerAuth": []
          }
        ],
        "tags": [
          "mp"
        ],
        "x-data-classification": "C1",
        "x-retention-class": "none",
        "x-smart-fasal-authorization": {
          "authentication": "mp",
          "capability": null,
          "denyByDefault": true,
          "purpose": null
        },
        "x-smart-fasal-problem-codes": [
          "AUTHENTICATION_REQUIRED",
          "AUTHORIZATION_DENIED",
          "DEPENDENCY_UNAVAILABLE",
          "AUTHORIZATION_VERSION_CHANGED",
          "MFA_REQUIRED"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://platform.api.smart-fasal.invalid"
    }
  ],
...
}